+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 /server/node_modules/
+/vue_project/node_modules/
