+++ vue_project/src/commonFunctions/index.js
@@ -1,0 +1,1 @@
+export * from './service';
