+++ vue_project/src/components/basketCard/basketCard.js
@@ -1,0 +1,8 @@
+Vue.component('basket-card',
+    {
+        template:
+            `<div class="basket-card">
+                <blue-button @click="$emit('view')">Закрыть</blue-button>
+                <slot></slot>
+            </div>`
+    });
