+++ vue_project/src/components/basketCard/index.js
@@ -1,0 +1,1 @@
+export * from './basketCard';
