+++ vue_project/src/components/basketItem/index.js
@@ -1,0 +1,1 @@
+export * from './basketItem';
