+++ vue_project/src/components/blueButton/blueButton.js
@@ -1,0 +1,8 @@
+Vue.component(
+    'blue-button',
+    {
+        template: `<button class="blue-button" @click="$emit('click', $event.target.click)">
+                      <slot></slot>
+                  </button>`
+    }
+);
