+++ vue_project/src/components/blueButton/index.js
@@ -1,0 +1,1 @@
+export * from './blueButton';
