+++ vue_project/src/components/blueSearch/blueSearch.js
@@ -1,0 +1,11 @@
+Vue.component(
+    'blue-search',
+    {
+        props: ['value'],
+        template:
+            `<input 
+            class="blue-input" 
+            v-bind:value="value" 
+            v-on:input="$emit('input', $event.target.value)">`
+    }
+);
