+++ vue_project/src/components/blueSearch/index.js
@@ -1,0 +1,1 @@
+export * from './blueSearch';
