+++ vue_project/src/components/goodsItem/index.js
@@ -1,0 +1,1 @@
+export * from './goodsItem'
