+++ vue_project/src/components/index.js
@@ -1,0 +1,5 @@
+export * from './basketItem';
+export * from './basketCard';
+export * from './blueSearch';
+export * from './blueButton';
+export * from './goodsItem';
