+++ vue_project/src/constants/constants.js
@@ -1,0 +1,6 @@
+export const baseAPIUrl = 'http://localhost:8000';
+export const APIUrls = {
+    getGoods: '/goods.json',
+    getBasket: '/basket-goods.json',
+    api: '/api'
+};
