+++ vue_project/webpack.config.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+const config = {
+    entry: './src/index.js',
+    output: {
+        filename: 'app.js',
+        path: path.resolve(__dirname, 'build')
+    }
+};
+
+module.exports = config;
